feat: stuck inputs and form into components, now just need to restructure architecture
